1. Spawning mechanic update
The start of the dreaded movement mechanic refactoring.
Step one is to refactor the spawning mechanic, and simplify the interactions with the grid manager and blocks.
Current revision: Shapes always try to spawn in the middle top of the board (5, 18), but if they can't (because they're too tall) they will automatically move down to fit the shape.
Tested with BigL, I, J, L, O, S, T and Z shapes